requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sse-starlette>=1.8.2
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
google-generativeai>=0.7.0
tavily-python>=0.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0
